added commands.sql

+++ web/proj/add_delete.php
@@ -1,4 +1,6 @@
-<?php include 'session.php';;
+<?php include 'session.php';
+
+$user_id = $_GET["user_id"];
 
 $action = $_GET["action"];
 $name = $_GET["name"];
@@ -9,12 +11,12 @@ $theDate = date("Y-m-d");
 //echo '<h1>'.$score.'</h1>';
 if ($action == 'delete'){
 	$session_id = $_GET["session_id"];
-	$stmt = $db->prepare('DELETE FROM game_session WHERE session_id = ' . $session_id);
+	$stmt = $db->prepare('DELETE FROM game_session WHERE session_id = ' . $session_id.'and user_id ='.$user_id);
 	$stmt->execute();
 }
 
 if ($action == 'add'){
-	$stmt = $db->prepare('INSERT INTO game_session (game_id, winner_name, score, date) VALUES  (:game_id, :winner, :score, :adate);');
+	$stmt = $db->prepare('INSERT INTO game_session (game_id, winner_name, score, date, user_id) VALUES  (:game_id, :winner, :score, :adate, :user_id);');
 	$stmt->bindValue(':game_id', $game_id, PDO::PARAM_INT);
 	$stmt->bindValue(':winner', $winner, PDO::PARAM_STR);
 	$stmt->bindValue(':score', $score, PDO::PARAM_STR);
